backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
mediapipe==0.10.18
opencv-python==4.8.1.78
numpy==1.26.2
python-dotenv==1.0.0